refactor: extract menu helpers and get_current_user to utils/handlers/helpers
- Move send_mentor_menu, send_trainee_menu, get_current_user from auth.py to utils/handlers/helpers.py
- Remove duplicate _send_mentor_menu_photo, _send_trainee_menu_photo from company.py
- Update imports in 7 handler files
- Set pytest testpaths to tests/unit so `pytest` runs only unit tests by default
