+++ pyproject.toml
@@ -1,5 +1,7 @@
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
+testpaths = ["tests/unit"]
 markers = [
+    "e2e: end-to-end tests (require Docker + Telegram accounts)",
     "order: set test execution order (e2e tests)",
 ]
